Add ability to restore NAND backups & inject NAND partitions
o NAND backups can be restored.
o CTRNAND & TWLN partitions can be injected separately, or all available partitions at once.
o These features can - when used in the wrong way - brick a 3DS. Users have to enter a unlock sequence before using to prevent accidential triggering.
o Changes in menu.h/.c to enable unlock sequence for 'dangerous' features.
o Change in hid.c required so that no event is triggered when a key is released. Also needed for unlock sequence.
o Change in draw.h/.c required so that the unlock sequence progress is properly displayed
o DecryptBuffer() renamed to CryptBuffer() (this is now also used for encrypting)
o  DecryptBufferInfo() renamed to CryptBufferInfo() (this is now also used for encrypting)

+++ source/draw.h
@@ -36,4 +36,4 @@ void DrawStringF(int x, int y, const char *format, ...);
 void DebugClear();
 void Debug(const char *format, ...);
 
-void ShowProgress(u32 current, u32 total);
+void ShowProgress(u64 current, u64 total);
